+++ apps/admin-ui/src/spa/recurring-reservations/application-rounds/[id]/review/AllocatedEventDataLoader.tsx
@@ -14,8 +14,8 @@ import { combineResults } from "@/common/util";
 import { useNotification } from "@/context/NotificationContext";
 import { APPLICATIONS_EVENTS_QUERY } from "./queries";
 import { FilterArguments } from "./Filters";
-import Loader from "../../Loader";
-import { More } from "../../lists/More";
+import Loader from "@/component/Loader";
+import { More } from "@/component/lists/More";
 import ApplicationEventsTable from "./ApplicationEventsTable";
 
 export type Sort = {
