+++ apps/admin-ui/src/spa/recurring-reservations/application-rounds/[id]/review/ApplicationsTable.tsx
